(NOT FINAL MILESTONE) Updated card display functionality
Updated display function for cards to allow for specified x and y placement of each card
Created a temporary value indicator showing the value of the first two cards
The value of a card is 10 if it is a Jack, Queen or King. I have yet to add the functionality of an Ace switching between 1 and 11.